requirements.txt
accelerate==0.30.1
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==3.7.1
astor==0.8.1
asyncer==0.0.2
attrdict==2.0.1
attrs==23.2.0
audioread==3.0.1
Authlib==1.3.1
auto_gptq==0.7.1
autodistill==0.1.28
autodistill-florence-2==0.1.1
Babel==2.15.0
backoff==2.2.1
bce-python-sdk==0.9.14
beautifulsoup4==4.12.3
bidict==0.23.1
bitsandbytes==0.43.1
blinker==1.8.2
boto3==1.34.130
botocore==1.34.130
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chainlit==1.1.305
chardet==4.0.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
cloudpickle==3.0.0
cohere==5.5.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cssutils==2.11.1
cuda-python==12.5.0
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.5.14
datasets==2.20.0
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
easyocr @ git+https://github.com/JaidedAI/EasyOCR.git@c999505ef6b43be1c4ee36aa04ad979175178352
einops==0.8.0
emoji==2.12.1
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.2
fastapi==0.110.3
fastavro==1.9.4
filelock==3.14.0
filetype==1.2.0
fire==0.6.0
flash-attn==2.6.2
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gekko==1.1.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.1
gradientai==1.13.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-status==1.62.2
grpcio-tools==1.64.1
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib-metadata==6.11.0
ipywidgets==8.1.3
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
Lazify==0.4.0
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.3.0
lightning-utilities==0.11.2
literalai==0.0.607
llvmlite==0.43.0
lmdb==1.4.1
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
mdurl==0.1.2
minijinja==2.0.1
mistralai==0.4.1
mlflow==2.13.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.34.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
optimum==1.20.0
ordered-set==4.1.0
orjson==3.10.5
packaging==23.2
paddleocr==2.7.3
paddlepaddle-gpu==2.6.1
pandas==2.2.2
parameterized==0.9.0
pdf2docx==0.5.8
peft==0.11.1
pillow==10.3.0
pinecone-client==3.2.2
pooch==1.8.2
portalocker==2.8.2
premailer==3.10.0
proto-plus==1.24.0.dev1
protobuf==4.25.3
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
PyJWT==2.8.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
python-bidi==0.5.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-socketio==5.11.3
pytorch-lightning==2.3.0
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
querystring-parser==1.2.4
rapidfuzz==3.9.3
rarfile==4.2
referencing==0.35.1
regex==2024.5.15
replicate==0.27.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.36
rouge==1.0.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.1
sacrebleu==2.4.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
sentencepiece==0.2.0
shapely==2.0.4
shtab==1.7.1
simple-websocket==1.0.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==2.0.30
sqlparse==0.5.0
starlette==0.37.2
streamlit==1.36.0
striprtf==0.0.26
supervision==0.22.0
sympy==1.12.1
syncer==2.0.3
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
text-generation==0.7.0
threadpoolctl==3.5.0
tifffile==2024.6.18
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.4
transformers==4.43.2
triton==2.3.1
trl==0.9.4
types-requests==2.32.0.20240602
typing-inspect==0.9.0
tyro==0.8.4
tzdata==2024.1
tzlocal==5.2
unstructured==0.14.9
unstructured-client==0.23.8
uptrace==1.24.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.25.0
validators==0.28.3
visualdl==2.5.3
watchdog==4.0.1
watchfiles==0.20.0
weaviate-client==3.26.5
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
